pyproject.toml
[project]
name = "bb-kb-mods"
version = "0.0.1"
authors = [
  { name="Harry Askham", email="harryaskham@gmail.com" },
]
description = "Programmatic edits to zitaotech BB keyboard models"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=1.26.0",
    "trimesh>=4.0.0",
    "matplotlib>=3.8.0",
    "pillow>=10.0.0",
    "click>=8.1.0",
    "scipy>=1.11.0",
    "networkx>=3.0",
]

[project.scripts]
bbkm = "bbkm.main:main"

[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = "."
module-name = ["bbkm"]
